+++ Libplanet.Tests/Action/ActionContextTest.cs
@@ -1,3 +1,4 @@
+using System.Collections.Immutable;
 using Libplanet.Action;
 using Xunit;
 
@@ -18,11 +19,27 @@ namespace Libplanet.Tests.Action
                 var context = new ActionContext(
                     signer: new Address("21744f4f08db23e044178dafb8273aeb5ebe6644"),
                     blockIndex: 1,
-                    previousStates: new AddressStateMap(),
+                    previousStates: new DumbAccountStateDelta(),
                     randomSeed: seed
                 );
                 IRandom random = context.Random;
                 Assert.Equal(expected, random.Next());
+            }
+        }
+
+        private class DumbAccountStateDelta : IAccountStateDelta
+        {
+            public IImmutableSet<Address> UpdatedAddresses =>
+                ImmutableHashSet<Address>.Empty;
+
+            public object GetState(Address address)
+            {
+                return null;
+            }
+
+            public IAccountStateDelta SetState(Address address, object state)
+            {
+                return this;
             }
         }
     }