+++ CHANGES.md
@@ -11,11 +11,15 @@ To be released.
     The corresponding parameter names on constructors and
     methods were also renamed too.
  -  Removed `IActionContext.To` property.
- -  Added `IAccountStateView` interface, and made `BlockChain<T>` implements it.
+ -  Added `IAccountStateView` interface.
     The interface purposes to provide a read-only view to account states.
- -  Added `IAccountStateDelta` interface.
+ -  Added `IAccountStateDelta` interface to replace `AddressStateMap`.
     The interface purposes to provide a read-write view to account states
     with maintaining `UpdatedAddresses` (so-called "dirty").
+ -  The type of `IActionContext.PreviousStates` property was changed from
+    `AddressStateMap` to `IAccountStateDelta`.
+ -  The return type of `IAction.Execute()` method was changed from
+    `AddressStateMap` to `IAccountStateDelta`.
  -  The type of `Peer.Urls` property was changed from `Uri` to `IPEndPoint`.
  -  Since we decided to depend on TURN ([RFC 5766]) and STUN ([RFC 5389]) to
     work around NAT so that `Peer`'s endpoints don't have to be multiple,